Refactor tests: Organize Issue #1 tests into proper structure
- Move helper functions to tests/conftest.py (shared fixtures)
- Move converter unit tests to tests/test_converter.py
- Move generator integration tests to tests/test_generator.py
- Remove duplicate tests/test_issue_1.py
- Move issue-1.md to docs/issues/resolved/
- Update TODO.md with current test coverage (29%, 16 passing)
- Clean up generated output/ directories

All Issue #1 tests retained with comments noting their origin.
Test coverage: 16 passed, 1 failed (known), 1 skipped
