+++ docs/issues/resolved/issue-1.md
@@ -1,6 +1,6 @@
 # Issue #1: Generated BDF fonts have incorrect character encodings
 
-**Status**: Identified and Reproduced
+**Status**: ✅ RESOLVED (merged 2025-11-29)
 **Severity**: High - Fonts do not render in applications
 **Component**: Font conversion pipeline (converter.py)
