+++ tests/test_converter.py
@@ -1,72 +1,30 @@
-"""
-Test for Issue #1: Generated BDF fonts have incorrect character encodings
+"""Tests for converter.py - Font conversion functions.
 
-This test suite verifies the fix for Issue #1, where BDF fonts generated by
-cp-font-gen had sequential encodings (1, 2, 3...) instead of correct Unicode
-codepoints (48, 49, 50... for digits '0', '1', '2'...).
-
-Test Structure:
-- Unit tests: Test fix_bdf_encodings() function directly
-- Integration tests: Test full generate_font() workflow
-- Regression tests: Document the original bug
+Includes unit tests for fix_bdf_encodings() and regression tests
+for the BDF encoding bug (Issue #1).
 """
 
 import os
-from pathlib import Path
 
 import pytest
 
 from cp_font_gen.converter import convert_to_bdf, fix_bdf_encodings, generate_subset_font
-from cp_font_gen.generator import generate_font
+
+from .conftest import create_test_bdf_with_wrong_encodings, extract_bdf_encodings
 
 
-# ============================================================================
-# Helper Functions
-# ============================================================================
+# =============================================================================
+# Unit Tests: fix_bdf_encodings() Function
+# =============================================================================
+# Originally from Issue #1 - tests for the encoding fix function
 
-def extract_bdf_encodings(bdf_path):
-    """Extract ENCODING values from a BDF file."""
-    encodings = []
-    with open(bdf_path, "r") as f:
-        for line in f:
-            if line.startswith("ENCODING "):
-                encoding = int(line.split()[1])
-                encodings.append(encoding)
-    return encodings
-
-
-def create_test_bdf_with_wrong_encodings(bdf_path, num_chars=10):
-    """Create a minimal BDF file with sequential encodings (1, 2, 3...) for testing."""
-    with open(bdf_path, "w") as f:
-        f.write("STARTFONT 2.1\n")
-        f.write("FONT -test-font\n")
-        f.write("SIZE 16 100 100\n")
-        f.write("FONTBOUNDINGBOX 10 16 0 0\n")
-        f.write("STARTPROPERTIES 1\n")
-        f.write("FONT_ASCENT 16\n")
-        f.write("ENDPROPERTIES\n")
-        f.write(f"CHARS {num_chars}\n")
-
-        # Write characters with WRONG sequential encodings (1, 2, 3...)
-        for i in range(num_chars):
-            f.write(f"STARTCHAR char{i}\n")
-            f.write(f"ENCODING {i + 1}\n")  # Wrong: should be 48-57 for digits
-            f.write("SWIDTH 500 0\n")
-            f.write("DWIDTH 10 0\n")
-            f.write("BBX 10 16 0 0\n")
-            f.write("BITMAP\n")
-            f.write("FF\n" * 16)  # Dummy bitmap
-            f.write("ENDCHAR\n")
-
-        f.write("ENDFONT\n")
-
-
-# ============================================================================
-# Unit Tests: Test fix_bdf_encodings() Function
-# ============================================================================
 
 def test_fix_bdf_encodings_corrects_sequential_to_unicode(tmp_path):
-    """Test that fix_bdf_encodings converts sequential encodings to Unicode codepoints."""
+    """Test that fix_bdf_encodings converts sequential encodings to Unicode codepoints.
+
+    Issue #1: BDF fonts had sequential encodings (1, 2, 3...) instead of
+    Unicode codepoints (48, 49, 50...). This tests the fix.
+    """
     # Create a BDF with wrong sequential encodings (1, 2, 3...)
     bdf_file = tmp_path / "test.bdf"
     create_test_bdf_with_wrong_encodings(bdf_file, num_chars=10)
@@ -91,7 +49,10 @@ def test_fix_bdf_encodings_corrects_sequential_to_unicode(tmp_path):
 
 
 def test_fix_bdf_encodings_preserves_correct_encodings(tmp_path):
-    """Test that fix_bdf_encodings doesn't change already-correct encodings."""
+    """Test that fix_bdf_encodings doesn't change already-correct encodings.
+
+    Issue #1 related: Ensure fix doesn't break fonts that already have correct encodings.
+    """
     # Create a BDF with CORRECT encodings (48-57)
     bdf_file = tmp_path / "test.bdf"
     with open(bdf_file, "w") as f:
@@ -128,7 +89,10 @@ def test_fix_bdf_encodings_preserves_correct_encodings(tmp_path):
 
 
 def test_fix_bdf_encodings_handles_various_characters(tmp_path):
-    """Test fix_bdf_encodings with different character types (letters, symbols)."""
+    """Test fix_bdf_encodings with different character types (letters, symbols).
+
+    Issue #1 related: Verify fix works for non-digit characters too.
+    """
     bdf_file = tmp_path / "test.bdf"
     create_test_bdf_with_wrong_encodings(bdf_file, num_chars=5)
 
@@ -143,85 +107,14 @@ def test_fix_bdf_encodings_handles_various_characters(tmp_path):
     assert encodings == expected
 
 
-# ============================================================================
-# Integration Tests: Test Full Workflow with generate_font()
-# ============================================================================
-
-def test_generate_font_produces_correct_encodings(tmp_path):
-    """Test that generate_font() workflow produces BDF with correct encodings (Issue #1 fix)."""
-    source_font = "/System/Library/Fonts/Helvetica.ttc"
-    if not os.path.exists(source_font):
-        pytest.skip("System font not available")
-
-    # Config for generating digits font
-    config = {
-        "source_font": source_font,
-        "sizes": [16],
-        "output": {
-            "formats": ["bdf"],
-            "font_family": "test-digits",
-        }
-    }
-
-    chars = set("0123456789")
-
-    # Generate font using full workflow
-    generated_files = generate_font(config, chars, tmp_path)
-
-    # Verify file was created
-    assert len(generated_files) > 0, "Should generate at least one file"
-
-    # Check the BDF file
-    bdf_file = tmp_path / "test-digits" / "test-digits-16pt.bdf"
-    assert bdf_file.exists(), f"BDF file should exist: {bdf_file}"
-
-    # Verify encodings are CORRECT (48-57 for '0'-'9')
-    encodings = extract_bdf_encodings(bdf_file)
-    expected = list(range(48, 58))  # 48-57 for '0'-'9'
-
-    assert encodings == expected, (
-        f"Generated BDF should have correct Unicode encodings (Issue #1 fix).\n"
-        f"Expected: {expected} (Unicode for '0'-'9')\n"
-        f"Got: {encodings}\n"
-        f"If this fails, the fix in generator.py may not be working."
-    )
-
-
-def test_generate_font_with_letters_has_correct_encodings(tmp_path):
-    """Test that fix works for non-digit characters too."""
-    source_font = "/System/Library/Fonts/Helvetica.ttc"
-    if not os.path.exists(source_font):
-        pytest.skip("System font not available")
-
-    config = {
-        "source_font": source_font,
-        "sizes": [16],
-        "output": {
-            "formats": ["bdf"],
-            "font_family": "test-abc",
-        }
-    }
-
-    chars = set("ABC")
-    generated_files = generate_font(config, chars, tmp_path)
-
-    assert len(generated_files) > 0
-
-    bdf_file = tmp_path / "test-abc" / "test-abc-16pt.bdf"
-    assert bdf_file.exists()
-
-    encodings = extract_bdf_encodings(bdf_file)
-    expected = [65, 66, 67]  # Unicode for 'A', 'B', 'C'
-    assert encodings == expected
-
-
-# ============================================================================
+# =============================================================================
 # Regression Tests: Document Original Bug
-# ============================================================================
+# =============================================================================
+# Originally from Issue #1 - documents the bug that was fixed
 
-def test_issue_1_bug_convert_to_bdf_produces_sequential_encodings(tmp_path):
-    """
-    REGRESSION TEST: Document the original Issue #1 bug.
+
+def test_issue_1_regression_convert_to_bdf_produces_sequential_encodings(tmp_path):
+    """REGRESSION TEST: Document the original Issue #1 bug.
 
     This test verifies that convert_to_bdf() alone (without fix_bdf_encodings)
     produces BDF files with sequential encodings (1, 2, 3...) instead of
@@ -275,13 +168,17 @@ def test_issue_1_bug_convert_to_bdf_produces_sequential_encodings(tmp_path):
     )
 
 
-def test_issue_1_integration_old_generated_fonts_have_bug():
-    """
-    REGRESSION TEST: Verify that old generated fonts (before fix) have the bug.
+def test_issue_1_regression_old_generated_fonts_have_bug():
+    """REGRESSION TEST: Verify that old generated fonts (before fix) have the bug.
 
     This checks if there are pre-existing generated fonts in output/ that
     demonstrate the original bug. Useful for before/after comparison.
+
+    Note: This test is expected to SKIP if no pre-existing fonts exist,
+    or PASS if fonts have been regenerated with the fix.
     """
+    from pathlib import Path
+
     output_dir = Path(__file__).parent.parent / "output" / "digits"
     bdf_file = output_dir / "digits-16pt.bdf"
 
@@ -310,8 +207,3 @@ def test_issue_1_integration_old_generated_fonts_have_bug():
             f"Expected either {expected_bug} (bug) or {expected_fixed} (fixed)\n"
             f"Got: {encodings}"
         )
-
-
-if __name__ == "__main__":
-    # Run tests directly
-    pytest.main([__file__, "-v"])
